requirements.txt
pyfunctional
matplotlib
beautifulsoup4
requests
pandas
numpy
sklearn
